# src/swagger/swagger.yml
openapi: 3.0.4
info:
  title: Swagger Petstore - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: https://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.12
...
servers:
  - url: https://petstore3.swagger.io/api/v3
...
security:
  - bearerAuth: []
paths:
  /api/countries:
    post:
      tags:
        - country
      summary: Create a new country
      description: Crea un nuevo país con nombre e imagen.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - image
              properties:
                name:
                  type: string
                  example: Colombia
                image:
                  type: string
                  example: https://example.com/images/colombia.png
      responses:
        '201':
          description: Country created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Country'
        '400':
          description: Invalid input
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
    get:
      tags:
        - country
      summary: Get all countries.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: Página actual para la paginación (opcional)
        - in: query
          name: limit
          schema:
            type: integer
          required: false
          description: Número de tickets por página (opcional)
      description: Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.
      operationId: getCountries
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Country'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Country'
        '400':
          description: Invalid tag value
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Country:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Colombia"
        image:
          type: string
          example: "https://example.com/images/colombia.png"
        isActive:
          type: boolean
          example: true
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: Ocurrió un error inesperado.
        code:
          type: integer
          example: 500
...